Fix conversions

+++ mock/MockIOperatingSystem.h
@@ -35,7 +35,7 @@ public:
     MOCK_METHOD(int, open, (const char* filename, int flags, int sflags), (override));
     MOCK_METHOD(int, stat, (const char* filename, struct stat* buf), (override));
     MOCK_METHOD(int, fstat, (int fd, struct stat* buf), (override));
-    MOCK_METHOD(int, lseek, (int fd, long offset, int origin), (override));
+    MOCK_METHOD(off_t, lseek, (int fd, off_t offset, int origin), (override));
     MOCK_METHOD(int, unlink, (const char* filename), (override));
     MOCK_METHOD(int, close, (int fd), (override));
     MOCK_METHOD(int, closeSocket, (SOCKET fd), (override));
